requirements.txt
asgiref==3.8.1
cryptography==3.4.8
diff-match-patch==20230430
Django==5.0.4
django-import-export==4.1.1
et-xmlfile==1.1.0
humanize==4.9.0
import-export==0.3.1
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
pillow==10.3.0
PyMySQL==1.1.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
